+++ src/utils/play.js
@@ -1,0 +1,11 @@
+import * as Tone from "tone";
+
+function play(song) {
+  //create a synth and connect it to the main output (your speakers)
+  const synth = new Tone.Synth().toDestination();
+  const now = Tone.now();
+
+  for (let i = 0; i < composition.length; i++) {
+    synth.triggerAttackRelease(composition[i], "4n", now + i);
+  }
+}
